move numpy_triple_part2 to numpy_triple

+++ benchmarks/verus/numpy_triple/files/strings_decode.rs
@@ -12,15 +12,15 @@ fn decode(a: Vec<Vec<u8>>, encoding: &str, errors: &str) -> (result: Vec<String>
     requires 
         encoding == "utf-8" || encoding == "ascii",
         errors == "strict" || errors == "ignore" || errors == "replace",
-        forall|i: int| 0 <= i < a.len() ==> a[i]@.len() >= 0,
+        forall|i: int| 0 <= i < a.len() as int ==> a[i]@.len() >= 0,
     ensures
-        result.len() == a.len(),
-        forall|i: int| 0 <= i < a.len() ==> (
+        result.len() as int == a.len() as int,
+        forall|i: int| 0 <= i < a.len() as int ==> (
             /* Basic well-formedness: decoded strings are valid */
             result[i]@.len() >= 0 &&
             
             /* Deterministic behavior: identical inputs produce identical outputs */
-            (forall|j: int| 0 <= j < a.len() && a[i]@ == a[j]@ ==> result[i]@ == result[j]@) &&
+            (forall|j: int| 0 <= j < a.len() as int && a[i]@ == a[j]@ ==> result[i]@ == result[j]@) &&
             
             /* Empty byte arrays decode to empty strings */
             (a[i]@.len() == 0 ==> result[i]@.len() == 0) &&
